% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicaser-package.R
\docType{package}
\name{epicaser-package}
\alias{epicaser}
\alias{epicaser-package}
\title{epicaser: Package to generate arrival of epi cases for queuing models, now includes hospitalisations}
\description{
A utility to generate cases and hospitalisations data.
}
\author{
\strong{Maintainer}: Jim Duggan \email{jim.duggan@universityofgalway.ie} (\href{https://orcid.org/0000-0002-7507-8617}{ORCID})

Authors:
\itemize{
  \item Jim Duggan \email{jim.duggan@universityofgalway.ie} (\href{https://orcid.org/0000-0002-7507-8617}{ORCID})
}

}
\keyword{internal}
