% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cohort_epi_cases.R
\name{generate_cohort_epi_cases}
\alias{generate_cohort_epi_cases}
\title{Takes a time series of cases and stratifies into groups
Groups and proportions are provided as inputs
Makes us of the package p2synthr}
\usage{
generate_cohort_epi_cases(
  dates,
  cases,
  groups = c("00-19", "20-39", "40-59", "60-79", "80-99", "100-110"),
  probs = c(0.15, 0.2, 0.2, 0.25, 0.19, 0.01),
  setSeed = F,
  seedValue = 100
)
}
\arguments{
\item{dates}{the times for each case}

\item{cases}{the case data}

\item{groups}{the age cohort groups}

\item{probs}{the probability of being in each group}

\item{setSeed}{allows setting of the seed for sampling}

\item{seedValue}{set the seed value to be used if needed}
}
\value{
tibble with results, including original value. Each stratified group has its own column.
}
\description{
\code{generate_cohort_epi_cases} Stratify epi cases by predefined cohorts
}
