% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_epi_case_list.R
\name{generate_epi_case_list}
\alias{generate_epi_case_list}
\title{Creates an individual list of cases based on the input cohorts. 
This structure will then form the basis for the hospitalisation structure}
\usage{
generate_epi_case_list(
  cohort_data,
  TS = FALSE,
  set_seed = FALSE,
  seed_val = 100
)
}
\arguments{
\item{cohort_data}{Cohort case data via SIR model and age cohort breakdown}

\item{TS}{Flag when true timestamps the synthetic data generated}

\item{set_seed}{allows setting of the seed for sampling}

\item{seed_val}{set the seed value to be used if needed}
}
\value{
res a tibble with day number, date, model output, cases, and available staff
}
\description{
\code{generate_epi_case_list} Create an epi case list based on model outputs
}
