% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_epi_cases_deterministic.R
\name{generate_epi_cases_deterministic}
\alias{generate_epi_cases_deterministic}
\title{Creates synthetic epi data using SIR model and Poisson/Negative Binomial
Staff absenteeism is proportional to the population attack rate (lambda)}
\usage{
generate_epi_cases_deterministic(
  N = 1e+05,
  R0 = 2,
  I0 = 1,
  RF = 0.5,
  sigma = 0.5,
  gamma = 0.5,
  start_time = 0,
  end_time = 100
)
}
\arguments{
\item{N}{the population size for the SIR model (default 100000)}

\item{R0}{the value for the reproduction number (default 2.0)}

\item{I0}{the initial number of people infected (default 1)}

\item{RF}{Reporting fraction for surveillance (default 0.5)}

\item{sigma}{Inverse of exposure delay}

\item{gamma}{Inverse of infectious delay}

\item{start_time}{start of simulation time (default 0)}

\item{end_time}{end of simulation time (default 100)}
}
\value{
res a tibble with day number and model output (cases)
}
\description{
\code{generate_epi_cases_deterministic} Create synthetic case data based on SIR model (Novel Pathogen)
}
