% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hospitalisation_data.R
\name{generate_hospitalisation_data}
\alias{generate_hospitalisation_data}
\title{Takes the case line list and updates with hospitalisation information}
\usage{
generate_hospitalisation_data(
  epi_cases,
  h_risk = dplyr::tibble(AgeL = c(0, 30, 70), AgeU = c(30, 70, 111), HRisk = c(0.01,
    0.05, 0.15)),
  time_to_admit_mean = 5.2,
  time_to_admit_sd = 2.1,
  h_time = dplyr::tibble(AgeL = c(0, 30, 70), AgeU = c(30, 70, 111), Mean = c(5, 10, 20),
    SD = c(1, 2, 5)),
  sources = c("Home", "Hospital", "Other"),
  source_prob = c(0.8, 0.1, 0.1),
  destinations = c("Home", "Hospital", "Death"),
  destination_prob = c(0.85, 0.13, 0.02),
  gender = c("M", "F"),
  gender_prob = c(0.49, 0.51),
  NL = FALSE
)
}
\arguments{
\item{epi_cases}{A tibble containing the case list of reported cases}

\item{h_risk}{A tibble that has the hospitalisation risk profiles for each age group}

\item{time_to_admit_mean}{The average time between diagnosis and admission to hospital}

\item{time_to_admit_sd}{The standard deviation for the average time}

\item{sources}{A vectors showing where the patient has come from before arriving at the hospital}

\item{destinations}{Where the patient goes after their length of stay}

\item{gender}{Gender of the patient}

\item{gender_prob}{Probability of each gender option}

\item{NL}{Netherlands flag, causes key fields to be renamed.}

\item{htime}{Length of stay distribution information (normal distribution)}

\item{sources_prob}{The probability of each sourcing option}
}
\value{
hosp_cases A tibble with all hospitalisation data
}
\description{
\code{generate_hospitalisation_data} Creates hospitalisation data that can be used for other modelling tasks
}
