% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_patient_pathways.R
\name{generate_patient_pathways}
\alias{generate_patient_pathways}
\title{Takes the case line list and updates with hospitalisation information}
\usage{
generate_patient_pathways(
  arrivals,
  pathway_logic = dplyr::tibble(Pathway_ID = c("P1", "P2", "P3", "P4", "P5", "P6"), AgeL
    = c(0, 0, 30, 30, 70, 70), AgeU = c(30, 30, 70, 70, 111, 111), Pathway =
    list(c("Ward", "Home"), c("Ward", "ICU", "Home"), c("Ward", "Home"), c("Ward", "ICU",
    "Home"), c("Ward", "Home"), c("Ward", "ICU", "Home")), Probabilities = c(0.95, 0.05,
    0.9, 0.1, 0.5, 0.5), Durations = list(c(1, 3 * 24), c(1, 2 * 24, 10 * 24), c(1, 5 *
    24), c(1, 7 * 24, 15 * 24), c(1, 10 * 24), c(1, 3 * 24, 25 * 24))),
  set_seed = TRUE,
  seed_val = 100
)
}
\arguments{
\item{arrivals}{A tibble containing patient arrivals at the hospital}

\item{pathway_logic}{A tibble that contains the pathway logic for each age group}

\item{set_seed}{allows setting of the seed for sampling}

\item{seed_val}{set the seed value to be used if needed}
}
\value{
A tibble with the pathways for each patient
}
\description{
\code{generate_patient_pathways} Generates arrival timestamped pathwats for all patients
}
