% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_patient_pathways_NL.R
\name{generate_patient_pathways_NL}
\alias{generate_patient_pathways_NL}
\title{Takes the case line list and updates with hospitalisation information}
\usage{
generate_patient_pathways_NL(
  arrivals,
  pathway_logic = dplyr::tibble(Pathway_ID = c("P1", "P2", "P3", "P4", "P5", "P6"), AgeL
    = c(0, 0, 30, 30, 70, 70), AgeU = c(30, 30, 70, 70, 111, 111), Internal_Pathway =
    list(c("Ward"), c("Ward", "ICU"), c("Ward"), c("Ward", "ICU"), c("Ward"), c("Ward",
    "ICU")), Final_Destination = c("Home", "Home", "Home", "Nursing home", "Home",
    "Home"), Probabilities = c(0.95, 0.05, 0.9, 0.1, 0.5, 0.5), Durations = list(c(3 *
    24), c(2 * 24, 10 * 24), c(5 * 24), c(7 * 24, 15 * 24), c(10 * 24), c(3 * 24, 25 *
    24))),
  set_seed = TRUE,
  seed_val = 100
)
}
\arguments{
\item{arrivals}{A tibble containing patient arrivals at the hospital}

\item{pathway_logic}{A tibble that contains the pathway logic for each age group}

\item{set_seed}{allows setting of the seed for sampling}

\item{seed_val}{set the seed value to be used if needed}
}
\value{
A tibble with the pathways for each patient
}
\description{
\code{generate_patient_pathways_NL} Generates arrival timestamped pathwats for all patients
}
