% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_timestamp.R
\name{generate_timestamp}
\alias{generate_timestamp}
\title{Generates a random timestamp for a specific date}
\usage{
generate_timestamp(date, hours = 8:22, mins = 0:59, secs = 0:59)
}
\arguments{
\item{date}{The date (Date object)}

\item{hours}{A range of hours to generate the hour}

\item{mins}{A range of minutes to generate the minute}

\item{secs}{A range of seconds to generate the second}
}
\value{
timestamp The combined timestamp to use in the patient record
}
\description{
\code{generate_timestamp} Generates timestamp for a specific date
}
