% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_epi_cases_stochastic.R
\name{generate_epi_cases_stochastic}
\alias{generate_epi_cases_stochastic}
\title{Creates synthetic epi data using SIR model and Poisson/Negative Binomial
Staff absenteeism is proportional to the population attack rate (lambda)}
\usage{
generate_epi_cases_stochastic(
  model_cases,
  start_day = "2025-03-01",
  seed = FALSE,
  seed_val = 100,
  Poisson = TRUE,
  nb_size = 10
)
}
\arguments{
\item{model_cases}{Output from the infectious disease model}

\item{start_day}{YYYY-MM-DD string format for start of simulation}

\item{seed}{True or False for using a seed (default FALSE)}

\item{seed_val}{value of seed (default 100)}

\item{Poisson}{Signifies whether to use the Poisson distribution synthetic data creation}

\item{nb_size}{size parameter for negative binomial distribution (default 10)}
}
\value{
res a tibble with day number, date, model output, cases, and available staff
}
\description{
\code{generate_epi_cases_stochastic} Create synthetic case data based on model input
}
