% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hospital_arrivals.R
\name{generate_hospital_arrivals}
\alias{generate_hospital_arrivals}
\title{Takes the case line list and updates with hospitalisation information}
\usage{
generate_hospital_arrivals(
  epi_cases,
  h_risk = dplyr::tibble(AgeL = c(0, 30, 70), AgeU = c(30, 70, 111), HRisk = c(0.03,
    0.08, 0.15)),
  time_to_admit_mean = 7,
  time_to_admit_sd = 2.1,
  sources = c("Home", "Other hospital", "Nursing home"),
  source_prob = c(0.55, 0.45, 0.05),
  gender = c("M", "F"),
  gender_prob = c(0.49, 0.51),
  set_seed = FALSE,
  seed_val = 100
)
}
\arguments{
\item{epi_cases}{A tibble containing the case list of reported cases}

\item{h_risk}{A tibble that has the hospitalisation risk profiles for each age group}

\item{time_to_admit_mean}{The average time between diagnosis and admission to hospital}

\item{time_to_admit_sd}{The standard deviation for the average time}

\item{sources}{A vectors showing where the patient has come from before arriving at the hospital}

\item{source_prob}{The probability of each sourcing option}

\item{gender}{Gender of the patient}

\item{gender_prob}{Probability of each gender option}

\item{set_seed}{allows setting of the seed for sampling}

\item{seed_val}{set the seed value to be used if needed}
}
\value{
hosp_cases A tibble with all hospitalisation data
}
\description{
\code{generate_hospital_arrivals} Generates arrival timestamps for patients
}
